Question prompt (#2)
* Fetch of 10 movies in the API

Fetch of 10 movies in the API, using a keyword, this is mandatory because the API won't let me fecth completly randomly. The fecth works only in the back and won't be return in the front for now

* Fetching three movie

Creating an array (I think?) to contain the three movie titles with a keyword, still couldn't manage to show it the HTML tho

* Creation of the question DB

Creation in raw code of all the question that the quiz can ask

* Creation of question and answers

For now, we have two arrays, one with the three false answers and one with the question, the right answer and one keyword to help finding better wrong answers. All of those are currently working separatly

* Keyword in DbFilm

The keyword from DbQuestion is now in DbFilm, but due to my being an idiot, this is the wrong version of DbFilm. Better late than never tho

* Questions almost finished (Back)

The code creates 2 arrays, one with the question, the right answer, and a keyword using for the fake answer. The second array has three fake answers looking like the real one

* Display of question and all answers

We display in basic HTML the quiz, we need to display the answers randomly